refactor: migrate programas to content collections
