+++ scripts/seeder/index.ts
@@ -2,6 +2,7 @@ import { seedDocentes } from './docentes';
 import { seedNoticias } from './noticias';
 import { seedAutoridades } from './autoridades';
 import { seedSustentaciones } from './sustentaciones';
+import { seedProgramas } from './programas';
 
 async function runSeeders() {
   console.log('🌱 Iniciando generación de datos (Seeders)...');
@@ -11,6 +12,7 @@ async function runSeeders() {
     await seedNoticias(25);
     await seedAutoridades();
     await seedSustentaciones(15);
+    await seedProgramas();
     
     console.log('✅ Todos los seeders se ejecutaron correctamente.');
   } catch (error) {
