+++ src/content/programas/doctorado-ciencias-ambientales.json
@@ -1,0 +1,22 @@
+{
+  "id": "doc-003",
+  "nombre": "Doctorado en Ciencias Ambientales",
+  "tipo": "doctorado",
+  "descripcion": "Programa interdisciplinario que forma investigadores en ciencias ambientales. Los doctorandos abordarán problemáticas ambientales desde perspectivas científicas, tecnológicas y de gestión.",
+  "descripcionCorta": "Investigación interdisciplinaria ambiental.",
+  "duracion": "3 años (6 semestres)",
+  "creditos": 96,
+  "modalidad": "presencial",
+  "estado": "activo",
+  "fechaInicio": "2025-03-15",
+  "slug": "doctorado-ciencias-ambientales",
+  "facultad": "Ingeniería",
+  "requisitos": [
+    "Grado de Magíster en ciencias o ingeniería ambiental",
+    "Proyecto de investigación doctoral",
+    "Experiencia en investigación (deseable)"
+  ],
+  "coordinador": "Dr. José Luis Vásquez Fernández",
+  "inversion": "S/. 24,000.00 (total)",
+  "imagen": "/images/programas/ambiental.jpg"
+}
