+++ src/data/programas.ts
@@ -1,299 +1,25 @@
-import type { Programa } from '@/types';
+import { getCollection, getEntry } from 'astro:content';
+import type { Programa } from '@/types/programas';
 
-export const programas: Programa[] = [
-  // MAESTRÍAS
-  {
-    id: 'mae-001',
-    nombre: 'Maestría en Gestión Pública',
-    tipo: 'maestria',
-    descripcion: 'Programa orientado a formar profesionales con competencias avanzadas en gestión y administración pública, capaces de liderar procesos de modernización del Estado y diseñar políticas públicas efectivas. Los egresados estarán preparados para ocupar cargos directivos en instituciones gubernamentales.',
-    descripcionCorta: 'Forma líderes en administración y modernización del Estado.',
-    duracion: '2 años (4 semestres)',
-    creditos: 72,
-    modalidad: 'semipresencial',
-    estado: 'activo',
-    fechaInicio: '2025-03-15',
-    slug: 'maestria-gestion-publica',
-    facultad: 'Ciencias Económicas y Administrativas',
-    requisitos: [
-      'Título profesional universitario',
-      'Experiencia laboral mínima de 2 años',
-      'Constancia de trabajo o declaración jurada',
-    ],
-    coordinador: 'Dr. Carlos Mendoza Ríos',
-    inversion: 'S/. 12,500.00 (total)',
-    imagen: '/images/programas/gestion-publica.jpg',
-    destacado: true,
-  },
-  {
-    id: 'mae-002',
-    nombre: 'Maestría en Derecho Civil y Comercial',
-    tipo: 'maestria',
-    descripcion: 'Programa que profundiza en el análisis del derecho civil y comercial, preparando profesionales para el ejercicio especializado del derecho en ámbitos contractuales, societarios y de negocios. Incluye análisis de jurisprudencia y casos prácticos.',
-    descripcionCorta: 'Especialización en derecho contractual, societario y de negocios.',
-    duracion: '2 años (4 semestres)',
-    creditos: 72,
-    modalidad: 'presencial',
-    estado: 'activo',
-    fechaInicio: '2025-03-15',
-    slug: 'maestria-derecho-civil-comercial',
-    facultad: 'Derecho y Ciencias Políticas',
-    requisitos: [
-      'Título de abogado',
-      'Colegiatura vigente',
-      'Constancia de habilitación profesional',
-    ],
-    coordinador: 'Dra. Patricia Vargas Lozano',
-    inversion: 'S/. 14,000.00 (total)',
-    imagen: '/images/programas/derecho.jpg',
-    destacado: true,
-  },
-  {
-    id: 'mae-003',
-    nombre: 'Maestría en Educación con mención en Docencia Universitaria',
-    tipo: 'maestria',
-    descripcion: 'Forma docentes universitarios con competencias pedagógicas avanzadas, capaces de diseñar, implementar y evaluar procesos de enseñanza-aprendizaje en educación superior. Incluye formación en investigación educativa y uso de tecnologías.',
-    descripcionCorta: 'Desarrolla competencias pedagógicas para la docencia universitaria.',
-    duracion: '2 años (4 semestres)',
-    creditos: 72,
-    modalidad: 'semipresencial',
-    estado: 'activo',
-    fechaInicio: '2025-03-15',
-    slug: 'maestria-educacion-docencia-universitaria',
-    facultad: 'Ciencias de la Educación',
-    requisitos: [
-      'Título profesional universitario',
-      'Experiencia docente mínima de 1 año (deseable)',
-    ],
-    coordinador: 'Dr. Roberto Sánchez Torres',
-    inversion: 'S/. 11,000.00 (total)',
-    imagen: '/images/programas/educacion.jpg',
-    destacado: true,
-  },
-  {
-    id: 'mae-004',
-    nombre: 'Maestría en Ingeniería Ambiental',
-    tipo: 'maestria',
-    descripcion: 'Programa enfocado en la gestión ambiental, evaluación de impacto ambiental y desarrollo sostenible. Los egresados podrán liderar proyectos ambientales y asesorar a organizaciones en cumplimiento de normativas ambientales.',
-    descripcionCorta: 'Gestión ambiental y desarrollo sostenible.',
-    duracion: '2 años (4 semestres)',
-    creditos: 72,
-    modalidad: 'presencial',
-    estado: 'activo',
-    fechaInicio: '2025-04-01',
-    slug: 'maestria-ingenieria-ambiental',
-    facultad: 'Ingeniería',
-    requisitos: [
-      'Título en ingeniería o ciencias afines',
-      'Conocimientos básicos de normativa ambiental',
-    ],
-    coordinador: 'Dr. Miguel Ángel Paredes Ruiz',
-    inversion: 'S/. 13,500.00 (total)',
-    imagen: '/images/programas/ambiental.jpg',
-  },
-  {
-    id: 'mae-005',
-    nombre: 'Maestría en Salud Pública',
-    tipo: 'maestria',
-    descripcion: 'Forma profesionales capaces de analizar, diseñar e implementar políticas y programas de salud pública. Incluye epidemiología, gestión de servicios de salud y promoción de la salud comunitaria.',
-    descripcionCorta: 'Políticas y programas de salud para la comunidad.',
-    duracion: '2 años (4 semestres)',
-    creditos: 72,
-    modalidad: 'semipresencial',
-    estado: 'activo',
-    fechaInicio: '2025-03-15',
-    slug: 'maestria-salud-publica',
-    facultad: 'Medicina Humana',
-    requisitos: [
-      'Título en ciencias de la salud',
-      'Experiencia en sector salud (deseable)',
-    ],
-    coordinador: 'Dra. Ana María Gutiérrez Flores',
-    inversion: 'S/. 15,000.00 (total)',
-    imagen: '/images/programas/salud.jpg',
-  },
+export async function getProgramas(): Promise<Programa[]> {
+  const collection = await getCollection('programas');
+  return collection.map(entry => ({
+    ...entry.data,
+    id: entry.id, // add the id manually as it is filename or content id
+  })) as Programa[];
+}
 
-  // DOCTORADOS
-  {
-    id: 'doc-001',
-    nombre: 'Doctorado en Derecho',
-    tipo: 'doctorado',
-    descripcion: 'Programa de alto nivel académico orientado a la investigación jurídica. Los doctorandos desarrollarán capacidades para producir conocimiento original en el campo del derecho y contribuir al debate académico nacional e internacional.',
-    descripcionCorta: 'Investigación jurídica de alto nivel.',
-    duracion: '3 años (6 semestres)',
-    creditos: 96,
-    modalidad: 'presencial',
-    estado: 'activo',
-    fechaInicio: '2025-03-15',
-    slug: 'doctorado-derecho',
-    facultad: 'Derecho y Ciencias Políticas',
-    requisitos: [
-      'Grado de Magíster en Derecho o áreas afines',
-      'Proyecto de investigación doctoral',
-      'Dominio de idioma extranjero (inglés o portugués)',
-    ],
-    coordinador: 'Dr. Fernando Castillo Mendoza',
-    inversion: 'S/. 25,000.00 (total)',
-    imagen: '/images/programas/derecho.jpg',
-    destacado: true,
-  },
-  {
-    id: 'doc-002',
-    nombre: 'Doctorado en Educación',
-    tipo: 'doctorado',
-    descripcion: 'Forma investigadores educativos capaces de generar conocimiento científico sobre procesos educativos. El programa enfatiza metodologías de investigación educativa y análisis crítico de políticas educativas.',
-    descripcionCorta: 'Investigación científica en educación.',
-    duracion: '3 años (6 semestres)',
-    creditos: 96,
-    modalidad: 'semipresencial',
-    estado: 'activo',
-    fechaInicio: '2025-04-01',
-    slug: 'doctorado-educacion',
-    facultad: 'Ciencias de la Educación',
-    requisitos: [
-      'Grado de Magíster en Educación o áreas afines',
-      'Proyecto de investigación doctoral',
-      'Publicación académica previa (deseable)',
-    ],
-    coordinador: 'Dra. Lucía Ramírez Delgado',
-    inversion: 'S/. 22,000.00 (total)',
-    imagen: '/images/programas/educacion.jpg',
-  },
-  {
-    id: 'doc-003',
-    nombre: 'Doctorado en Ciencias Ambientales',
-    tipo: 'doctorado',
-    descripcion: 'Programa interdisciplinario que forma investigadores en ciencias ambientales. Los doctorandos abordarán problemáticas ambientales desde perspectivas científicas, tecnológicas y de gestión.',
-    descripcionCorta: 'Investigación interdisciplinaria ambiental.',
-    duracion: '3 años (6 semestres)',
-    creditos: 96,
-    modalidad: 'presencial',
-    estado: 'activo',
-    fechaInicio: '2025-03-15',
-    slug: 'doctorado-ciencias-ambientales',
-    facultad: 'Ingeniería',
-    requisitos: [
-      'Grado de Magíster en ciencias o ingeniería ambiental',
-      'Proyecto de investigación doctoral',
-      'Experiencia en investigación (deseable)',
-    ],
-    coordinador: 'Dr. José Luis Vásquez Fernández',
-    inversion: 'S/. 24,000.00 (total)',
-    imagen: '/images/programas/ambiental.jpg',
-  },
+export async function getProgramaBySlug(slug: string): Promise<Programa | undefined> {
+  // Try to find it. In Astro data collections with JSON, the ID is often the filename without extension.
+  // Wait, the id parameter for getEntry is the collection entry id (filename).
+  // If the files are named slug.json, then the id IS the slug.
+  const entry = await getEntry('programas', slug);
+  if (!entry) return undefined;
+  return { id: entry.id, ...entry.data } as Programa;
+}
 
-  // DIPLOMADOS (Formación Continua)
-  {
-    id: 'dip-001',
-    nombre: 'Diplomado en Gestión de Proyectos de Inversión Pública',
-    tipo: 'diplomado',
-    descripcion: 'Capacita profesionales en la formulación, evaluación y gestión de proyectos de inversión pública bajo el marco del Sistema Nacional de Programación Multianual y Gestión de Inversiones (Invierte.pe).',
-    descripcionCorta: 'Formulación y gestión de proyectos bajo Invierte.pe.',
-    duracion: '6 meses',
-    creditos: 24,
-    modalidad: 'virtual',
-    estado: 'activo',
-    fechaInicio: '2025-02-15',
-    slug: 'diplomado-proyectos-inversion-publica',
-    facultad: 'Ciencias Económicas y Administrativas',
-    requisitos: [
-      'Título profesional o bachiller universitario',
-      'Acceso a computadora e internet',
-    ],
-    coordinador: 'Mg. Ricardo Flores Pérez',
-    inversion: 'S/. 2,800.00 (total)',
-    imagen: '/images/programas/proyectos.jpg',
-  },
-  {
-    id: 'dip-002',
-    nombre: 'Diplomado en Derecho Administrativo',
-    tipo: 'diplomado',
-    descripcion: 'Proporciona conocimientos actualizados sobre el derecho administrativo peruano, procedimientos administrativos, contrataciones del Estado y responsabilidad administrativa.',
-    descripcionCorta: 'Actualización en derecho administrativo peruano.',
-    duracion: '4 meses',
-    creditos: 20,
-    modalidad: 'semipresencial',
-    estado: 'activo',
-    fechaInicio: '2025-02-01',
-    slug: 'diplomado-derecho-administrativo',
-    facultad: 'Derecho y Ciencias Políticas',
-    requisitos: [
-      'Bachiller o título profesional',
-    ],
-    coordinador: 'Mg. Sandra López Medina',
-    inversion: 'S/. 2,200.00 (total)',
-    imagen: '/images/programas/derecho.jpg',
-  },
-  {
-    id: 'dip-003',
-    nombre: 'Diplomado en Gestión Educativa',
-    tipo: 'diplomado',
-    descripcion: 'Desarrolla competencias de liderazgo y gestión para directivos de instituciones educativas. Incluye planificación estratégica, gestión del talento humano y mejora continua.',
-    descripcionCorta: 'Liderazgo y gestión para directivos educativos.',
-    duracion: '5 meses',
-    creditos: 22,
-    modalidad: 'virtual',
-    estado: 'activo',
-    fechaInicio: '2025-03-01',
-    slug: 'diplomado-gestion-educativa',
-    facultad: 'Ciencias de la Educación',
-    requisitos: [
-      'Título profesional en educación o áreas afines',
-      'Experiencia en gestión educativa (deseable)',
-    ],
-    coordinador: 'Mg. Elena Torres Campos',
-    inversion: 'S/. 2,500.00 (total)',
-    imagen: '/images/programas/educacion.jpg',
-  },
-
-  // CURSOS
-  {
-    id: 'cur-001',
-    nombre: 'Curso de Actualización en Normas APA 7ma Edición',
-    tipo: 'curso',
-    descripcion: 'Curso práctico sobre el uso de las normas APA en su séptima edición para la redacción de trabajos académicos, tesis y artículos científicos.',
-    descripcionCorta: 'Redacción académica con normas APA.',
-    duracion: '4 semanas',
-    creditos: 4,
-    modalidad: 'virtual',
-    estado: 'activo',
-    fechaInicio: '2025-02-10',
-    slug: 'curso-normas-apa',
-    facultad: 'Escuela de Postgrado',
-    requisitos: [
-      'Estudiante o egresado universitario',
-    ],
-    coordinador: 'Mg. Carmen Díaz Ruiz',
-    inversion: 'S/. 180.00 (total)',
-    imagen: '/images/programas/metodologia.jpg',
-  },
-  {
-    id: 'cur-002',
-    nombre: 'Curso de Metodología de la Investigación Científica',
-    tipo: 'curso',
-    descripcion: 'Fundamentos de la investigación científica, diseño de proyectos de investigación, métodos cuantitativos y cualitativos, y elaboración de instrumentos de recolección de datos.',
-    descripcionCorta: 'Fundamentos para la investigación académica.',
-    duracion: '6 semanas',
-    creditos: 6,
-    modalidad: 'virtual',
-    estado: 'activo',
-    fechaInicio: '2025-02-17',
-    slug: 'curso-metodologia-investigacion',
-    facultad: 'Escuela de Postgrado',
-    requisitos: [
-      'Bachiller o estudiante de postgrado',
-    ],
-    coordinador: 'Dr. Alberto Ramos García',
-    inversion: 'S/. 350.00 (total)',
-    imagen: '/images/programas/metodologia.jpg',
-  },
-];
-
-// Helpers para filtrar programas
-export const getMaestrias = () => programas.filter(p => p.tipo === 'maestria');
-export const getDoctorados = () => programas.filter(p => p.tipo === 'doctorado');
-export const getDiplomados = () => programas.filter(p => p.tipo === 'diplomado');
-export const getCursos = () => programas.filter(p => p.tipo === 'curso');
-export const getProgramasActivos = () => programas.filter(p => p.estado === 'activo');
-export const getProgramaBySlug = (slug: string) => programas.find(p => p.slug === slug);
-export const getProgramaById = (id: string) => programas.find(p => p.slug === id || p.id === id);
+export async function getProgramaById(id: string): Promise<Programa | undefined> {
+  // If the file is 'mae-001.json' then id is 'mae-001'. But our files are named by slug, so 'id' field is inside data.
+  const all = await getProgramas();
+  return all.find(p => p.id === id);
+}
