+++ src/pages/docentes/[id].astro
@@ -3,6 +3,8 @@ import { getCollection } from 'astro:content';
 import Layout from '../../layouts/Layout.astro';
 import Navbar from '../../features/navigation/components/Navbar.astro';
 import { DocenteDetail } from '../../features/docentes';
+import { getProgramaById } from '../../data/programas';
+import type { Programa } from '../../types/programas';
 import { ErrorBoundary } from '../../components/ui/error-boundary';
 
 export async function getStaticPaths() {
@@ -33,6 +35,11 @@ const gradoLabels: Record<string, string> = {
 };
 
 const nombreCompleto = `${gradoLabels[docente.grado] || ''} ${docente.nombres} ${docente.apellidos}`;
+
+const programasDocenteUnresolved = await Promise.all(
+  (docente.programas || []).map(id => getProgramaById(id))
+);
+const programasDocente = programasDocenteUnresolved.filter(Boolean) as Programa[];
 ---
 
 <Layout 
@@ -44,7 +51,7 @@ const nombreCompleto = `${gradoLabels[docente.grado] || ''} ${docente.nombres} $
   <Navbar />
   <main class="min-h-screen">
     <ErrorBoundary client:load>
-      <DocenteDetail client:load docente={docente} />
+      <DocenteDetail client:load docente={docente} programasDocente={programasDocente} />
     </ErrorBoundary>
   </main>
 </Layout>
